data_ingestor/requirements.txt
Flask==2.2.2
google-cloud-pubsub==2.13.4
websockets==10.3
gunicorn==20.1.0
